restructure LandingPage / App / Login
